WIP. First round of JPA additions.

+++ kp-library-api/src/main/java/de/kaiserpfalzedv/office/library/api/HasRecord.java
@@ -26,7 +26,7 @@ import java.time.OffsetDateTime;
  * @since 1.0.0  2023-01-15
  */
 public interface HasRecord {
-    Long getVersion();
+    Integer getVersion();
     OffsetDateTime getCreated();
     OffsetDateTime getModified();
     OffsetDateTime getRevisioned();
